Add `Drop` impl for `PoolGuard` and more Pool public APIs
